+++ src/Shared/ApiModels.fs
@@ -1,5 +1,14 @@
 ﻿module ApiModels
 
+// Types defined here are shared between the server and the client.
+// Since we currently use native js deserialization with casting on the client side (zero overhead),
+// it's important to use only simple or compatible types, to avoid nasty surprises.
+//
+// For example, int64 cannot be used transparently this way (see https://fable.io/docs/dotnet/numbers.html)
+// and will lead to runtime errors when Fable uses them as its internal int64 representation but the value is really
+// just a js Number coming straight from json.
+// A simple solution is to use float (double in C#) instead, which translate directly to js Number...
+
 type GenericError = {
     Error: string
 }
@@ -8,13 +17,13 @@ type FlatScrobble = {
     Artist: string
     Album: string
     Track: string
-    Timestamp: int64
+    Timestamp: float // int64
 }
 
 type UserName = UserName of string
 
 type GetResumeTimestampResponse = {
-    ResumeFrom: int64
+    ResumeFrom: float // int64
 }
 
 type InsertScrobblesResponse = {
@@ -26,7 +35,7 @@ type InsertScrobblesResponse = {
 }
 
 type GetChartRequestOptions = {
-    PageToken: int64 option
+    PageToken: float option // int64
     PageSize: int option
 }
 
@@ -35,7 +44,7 @@ type GetChartDataResponse = {
     Colors: string[]
     Texts: string[]
     // Paging properties:
-    NextPageToken: int64
+    NextPageToken: float // int64
     TotalCount: int
 }
